requirements.txt
flask
python-pptx
openai
matplotlib